+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/layers/convolution/Deconvolution2DLayer.java
@@ -171,8 +171,9 @@ public class Deconvolution2DLayer extends ConvolutionLayer {
                     + " " + layerId());
         }
 
-        long inDepth = weights.size(0);
-        long outDepth = weights.size(1);
+        // FIXME: int cast
+        int inDepth = (int) weights.size(0);
+        int outDepth = (int) weights.size(1);
 
         if (input.size(1) != inDepth && input.size(3) == inDepth) {
             input = input.permute(0, 3, 1, 2);
@@ -186,8 +187,8 @@ public class Deconvolution2DLayer extends ConvolutionLayer {
                     + Arrays.toString(input.shape()) + "; expected" + " input channels = " + inDepth + ") "
                     + layerId());
         }
-        long kH = weights.size(2);
-        long kW = weights.size(3);
+        int kH = (int) weights.size(2);
+        int kW = (int) weights.size(3);
 
         int[] dilation = layerConf().getDilation();
         int[] kernel = layerConf().getKernelSize();
@@ -215,7 +216,7 @@ public class Deconvolution2DLayer extends ConvolutionLayer {
         int sameMode = (convolutionMode == ConvolutionMode.Same) ? 1 : 0;
 
         int[] args = new int[] {
-                (int)kH, (int)kW, strides[0], strides[1],
+                kH, kW, strides[0], strides[1],
                 pad[0], pad[1], dilation[0], dilation[1], sameMode, 0   //Last arg: 0 for nchw
         };
